backend/requirements.txt
# LLM & Vector DB
google-generativeai
pinecone
openai

# FastAPI Web Server
fastapi
uvicorn
python-multipart

# ML & Image/Text Processing
torch
# Microsoft BioViL + Vision-Language Toolbox (import path: health_multimodal)
hi-ml-multimodal
konlpy
scikit-learn
numpy
beautifulsoup4

# Data Handling & Visualization
pandas
matplotlib
seaborn

# UI & Utilities
rich
python-dotenv
Pillow
